Major cleanup
Move packages around
Tons of comments added
Couple of methods added and/or moved
