+++ src/main/java/gr8pefish/ironbackpacks/core/RecipesIronBackpacks.java
@@ -2,7 +2,7 @@ package gr8pefish.ironbackpacks.core;
 
 import gr8pefish.ironbackpacks.ConfigHandler;
 import gr8pefish.ironbackpacks.IronBackpacks;
-import gr8pefish.ironbackpacks.api.variant.BackpackSpecialty;
+import gr8pefish.ironbackpacks.api.backpack.variant.BackpackSpecialty;
 import gr8pefish.ironbackpacks.api.IronBackpacksAPI;
 import gr8pefish.ironbackpacks.core.recipe.BackpackTierRecipe;
 import net.minecraft.init.Blocks;
