+++ src/main/java/gr8pefish/ironbackpacks/item/ItemBackpack.java
@@ -6,8 +6,8 @@ import gr8pefish.ironbackpacks.api.*;
 import gr8pefish.ironbackpacks.api.backpack.BackpackInfo;
 import gr8pefish.ironbackpacks.api.backpack.IBackpack;
 import gr8pefish.ironbackpacks.api.upgrade.BackpackUpgrade;
-import gr8pefish.ironbackpacks.api.variant.BackpackSpecialty;
-import gr8pefish.ironbackpacks.api.variant.BackpackType;
+import gr8pefish.ironbackpacks.api.backpack.variant.BackpackSpecialty;
+import gr8pefish.ironbackpacks.api.backpack.variant.BackpackType;
 import gr8pefish.ironbackpacks.capabilities.BackpackInvImpl;
 import gr8pefish.ironbackpacks.core.RegistrarIronBackpacks;
 import gr8pefish.ironbackpacks.network.GuiHandler;
